add manual setup of k3s with Flux on Compute Engine
This currently is not fully working because we lost the metadata,
and we need this to hook up all of flux. If/when Google Build is working
again I should be able to generate the images with the addition of
cgroups version 2 and everything will come together. For now you
have to run the install steps manually (sorry).

Signed-off-by: vsoch <vsoch@users.noreply.github.com>