tested flyte - like it as well.
I want to next try creating a flux operator plugin. Flyte is nice because
it feels like a production tool, and it provides handles to a lot of the
kubernetes workflow tools we want (e.g., a few of us like KubeFlow).
My one concern is that the complexity might be overkill - e.g., it
is a LOT to deploy a cluster just to run these jobs. But in that the
final user interaction is writing a python script, and we do not have
to do that compile and DSL nonsense like with KubeFlow pipelines? That
seems a lot easier.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>