add resource test
we want to test if setting the resource limits/requests actually
limits what flux sees. We saw this once on a GCP run cluster.
It does not appear to work for us (at least I could not
reproduce it, with or without the affinity rule, and
with or without the cpu + memory.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>