add wip performance run
here we are trying to see how (when we set a limit/request for resources)
the mapping of cpu actually happens. Given the commonality of an OOMKIlled
I was only able to add one more pod, meaning one node had two pods on
it for a flux cluster. Hopefully that is enough to start to see something.
The other interesting observation is that with this Guaranteed status
and limits, it was not even possible to dnf install stuff. So I hugely
question the real world utility of setting these.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>