updates: design for run9
We cannot use the success policy, so we need to monitor only the 0 index
and have the others sleep, and then terminate the group. We also will
do better off waiting for index0 to finish and then doing one call to
get the log. In watching, the long breaks seem to be waiting for groups
to Terminate, which is not super quick. I have not seen clogging yet and
saw one small size (iters 5) run through completion. I also want to note
that the submit to end time is not entirely accurate because we have to
account for the time to finish submit and THEN start monitoring with
multiprocess. That said, we can probably time that entire loop and then
subtract it.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>